requirements.txt
flask==2.0.1
qrcode==7.3.1
pillow==10.4.0
werkzeug==2.0.3
python-dotenv==1.0.1
gunicorn==22.0.0